working code used to pre-process omol data into adios files.
omol25 differs from omat24 in the following ways:

- omat has 5300 aselmdb files whereas omol has 80
- omat files have different number of samples in each file, whereas all
files in omol have the same number of samples.
- raw omat data is under 500 GB in size whereas raw omol data is around
1 TB.
- some of the aselmdb files of omat seem to be corrupted. no such issues
for omol.

Consequently, as omol has 80 data files, we do not distribute data files to
ranks. Instead, for every data file, we distribute the approximately 1+
million samples amongst all available ranks for greater parallelism.

On the contrary, for omat24, we distribute data files amongst available
ranks.
