+++ examples/open_molecules_2025/train.py
@@ -24,10 +24,6 @@ spec.loader.exec_module(mod)
 
 ###############################################################################################
 
-import numpy as np
-
-import random
-
 import torch
 
 # FIX random seed
@@ -56,9 +52,7 @@ from hydragnn.preprocess.graph_samples_checks_and_updates import (
     gather_deg,
 )
 from hydragnn.preprocess.load_data import split_dataset
-
 import hydragnn.utils.profiling_and_tracing.tracer as tr
-
 from hydragnn.utils.print.print_utils import iterate_tqdm, log
 
 try:
@@ -68,6 +62,7 @@ except ImportError:
 
 import subprocess
 from hydragnn.utils.distributed import nsplit
+from omol25 import OMol2025
 
 ## FIMME
 torch.backends.cudnn.enabled = False
@@ -78,241 +73,6 @@ import glob
 
 def info(*args, logtype="info", sep=" "):
     getattr(logging, logtype)(sep.join(map(str, args)))
-
-
-# transform_coordinates = Spherical(norm=False, cat=False)
-transform_coordinates = LocalCartesian(norm=False, cat=False)
-# transform_coordinates = Distance(norm=False, cat=False)
-
-transform_coordinates_pbc = PBCLocalCartesian(norm=False, cat=False)
-# transform_coordinates_pbc = PBCDistance(norm=False, cat=False)
-
-dataset_names = [
-    "rattled-300-subsampled",
-]
-
-
-class OMol2025(AbstractBaseDataset):
-    def __init__(
-        self,
-        dirpath,
-        config,
-        data_type,
-        graphgps_transform=None,
-        energy_per_atom=True,
-        dist=False,
-    ):
-        super().__init__()
-
-        assert (data_type == "train") or (
-            data_type == "val"
-        ), "data_type must be a string either equal to 'train' or to 'val'"
-
-        self.config = config
-        self.radius = config["NeuralNetwork"]["Architecture"]["radius"]
-        self.max_neighbours = config["NeuralNetwork"]["Architecture"]["max_neighbours"]
-
-        self.data_path = os.path.join(dirpath, data_type)
-        self.energy_per_atom = energy_per_atom
-
-        self.radius_graph = RadiusGraph(
-            self.radius, loop=False, max_num_neighbors=self.max_neighbours
-        )
-        self.radius_graph_pbc = RadiusGraphPBC(
-            self.radius, loop=False, max_num_neighbors=self.max_neighbours
-        )
-
-        self.graphgps_transform = graphgps_transform
-
-        config_kwargs = {}  # see tutorial on additional configuration
-
-        # Threshold for atomic forces in eV/angstrom
-        self.forces_norm_threshold = 1000.0
-
-        self.dist = dist
-        if self.dist:
-            assert torch.distributed.is_initialized()
-            self.world_size = torch.distributed.get_world_size()
-            self.rank = torch.distributed.get_rank()
-
-        ## Parallelizing over data files for training data. For val set, we parallelize over each molecules.
-        if data_type == "train":
-            total_file_list = glob.glob(
-                os.path.join(dirpath, data_type, "*.aselmdb"), recursive=True
-            )
-            rx = list(nsplit(total_file_list, self.world_size))[self.rank]
-            print(self.rank, "total:", len(total_file_list), "local:", len(rx))
-            dataset_list = rx
-        else:
-            dataset_list = dataset_names
-
-        torch.distributed.barrier()
-
-        for dataset_index, dataname in enumerate(dataset_list):
-
-            # print(f"Rank {self.rank} reading {data_type}/{dataname} ... ", flush=True)
-
-            dataset = AseDBDataset(
-                config=dict(
-                    src=os.path.join(dirpath, data_type, dataname), **config_kwargs
-                )
-            )
-            # print(self.rank, "dataname:", dataname, "total:", dataset.num_samples)
-
-            if data_type == "train":
-                rx = list(range(dataset.num_samples))
-            else:
-                rx = list(nsplit(list(range(dataset.num_samples)), self.world_size))[
-                    self.rank
-                ]
-                print(
-                    self.rank,
-                    "dataname:",
-                    dataname,
-                    "total:",
-                    dataset.num_samples,
-                    "local:",
-                    len(rx),
-                )
-
-            for index in iterate_tqdm(
-                rx,
-                verbosity_level=2,
-                desc=f"Rank{self.rank} Dataset {dataset_index}/{len(dataset_list)}",
-            ):
-                try:
-                    pos = torch.tensor(
-                        dataset.get_atoms(index).get_positions(), dtype=torch.float32
-                    )
-                    natoms = torch.IntTensor([pos.shape[0]])
-                    atomic_numbers = torch.tensor(
-                        dataset.get_atoms(index).get_atomic_numbers(),
-                        dtype=torch.float32,
-                    ).unsqueeze(1)
-                    energy = torch.tensor(
-                        dataset.get_atoms(index).get_total_energy(), dtype=torch.float32
-                    ).unsqueeze(0)
-                    energy_per_atom = energy.detach().clone() / natoms
-                    forces = torch.tensor(
-                        dataset.get_atoms(index).get_forces(), dtype=torch.float32
-                    )
-                    chemical_formula = dataset.get_atoms(index).get_chemical_formula()
-
-                    cell = None
-                    try:
-                        cell = torch.tensor(
-                            dataset.get_atoms(index).get_cell(), dtype=torch.float32
-                        ).view(3, 3)
-                    except:
-                        print(
-                            f"Atomic structure {chemical_formula} does not have cell",
-                            flush=True,
-                        )
-
-                    pbc = None
-                    try:
-                        pbc = pbc_as_tensor(dataset.get_atoms(index).get_pbc())
-                    except:
-                        print(
-                            f"Atomic structure {chemical_formula} does not have pbc",
-                            flush=True,
-                        )
-
-                    # If either cell or pbc were not read, we set to defaults which are not none.
-                    if cell is None or pbc is None:
-                        cell = torch.eye(3, dtype=torch.float32)
-                        pbc = torch.tensor([False, False, False], dtype=torch.bool)
-
-                    x = torch.cat([atomic_numbers, pos, forces], dim=1)
-
-                    # Calculate chemical composition
-                    atomic_number_list = atomic_numbers.tolist()
-                    assert len(atomic_number_list) == natoms
-                    ## 118: number of atoms in the periodic table
-                    hist, _ = np.histogram(atomic_number_list, bins=range(1, 118 + 2))
-                    chemical_composition = (
-                        torch.tensor(hist).unsqueeze(1).to(torch.float32)
-                    )
-
-                    data_object = Data(
-                        dataset_name="omat24",
-                        natoms=natoms,
-                        pos=pos,
-                        cell=cell,
-                        pbc=pbc,
-                        edge_index=None,
-                        edge_attr=None,
-                        atomic_numbers=atomic_numbers,
-                        chemical_composition=chemical_composition,
-                        smiles_string=None,
-                        x=x,
-                        energy=energy,
-                        energy_per_atom=energy_per_atom,
-                        forces=forces,
-                    )
-
-                    if self.energy_per_atom:
-                        data_object.y = data_object.energy_per_atom
-                    else:
-                        data_object.y = data_object.energy
-
-                    if data_object.pbc.any():
-                        try:
-                            data_object = self.radius_graph_pbc(data_object)
-                            data_object = transform_coordinates_pbc(data_object)
-                        except:
-                            print(
-                                f"Structure could not successfully apply one or both of the pbc radius graph and positional transform",
-                                flush=True,
-                            )
-                            data_object = self.radius_graph(data_object)
-                            data_object = transform_coordinates(data_object)
-                    else:
-                        data_object = self.radius_graph(data_object)
-                        data_object = transform_coordinates(data_object)
-
-                    # Default edge_shifts for when radius_graph_pbc is not activated
-                    if not hasattr(data_object, "edge_shifts"):
-                        data_object.edge_shifts = torch.zeros(
-                            (data_object.edge_index.size(1), 3), dtype=torch.float32
-                        )
-
-                    # FIXME: PBC from bool --> int32 to be accepted by ADIOS
-                    data_object.pbc = data_object.pbc.int()
-
-                    # LPE
-                    if self.graphgps_transform is not None:
-                        data_object = self.graphgps_transform(data_object)
-
-                    if self.check_forces_values(data_object.forces):
-                        self.dataset.append(data_object)
-                    else:
-                        print(
-                            f"L2-norm of force tensor is {data_object.forces.norm()} and exceeds threshold {self.forces_norm_threshold} - atomistic structure: {chemical_formula}",
-                            flush=True,
-                        )
-
-                except Exception as e:
-                    print(f"Rank {self.rank} reading - exception: ", e)
-
-        print(self.rank, "Done. Wait others.", flush=True)
-        torch.distributed.barrier()
-
-        random.shuffle(self.dataset)
-
-    def check_forces_values(self, forces):
-
-        # Calculate the L2 norm for each row
-        norms = torch.norm(forces, p=2, dim=1)
-        # Check if all norms are less than the threshold
-
-        return torch.all(norms < self.forces_norm_threshold).item()
-
-    def len(self):
-        return len(self.dataset)
-
-    def get(self, idx):
-        return self.dataset[idx]
 
 
 if __name__ == "__main__":
@@ -370,6 +130,7 @@ if __name__ == "__main__":
     node_feature_dims = [1, 3, 3]
     dirpwd = os.path.dirname(os.path.abspath(__file__))
     datadir = os.path.join(dirpwd, "dataset")
+    
     ##################################################################################################################
     input_filename = os.path.join(dirpwd, args.inputfile)
     ##################################################################################################################
@@ -412,13 +173,13 @@ if __name__ == "__main__":
         datefmt="%H:%M:%S",
     )
 
-    log_name = "OMat24" if args.log is None else args.log
+    log_name = "OMol25" if args.log is None else args.log
     hydragnn.utils.print.setup_log(log_name)
     writer = hydragnn.utils.model.get_summary_writer(log_name)
 
     log("Command: {0}\n".format(" ".join([x for x in sys.argv])), rank=0)
 
-    modelname = "OMat24" if args.modelname is None else args.modelname
+    modelname = "OMol25" if args.modelname is None else args.modelname
     if args.preonly:
         ## local data
         trainset = OMol2025(
@@ -449,15 +210,14 @@ if __name__ == "__main__":
         ## Need as a list
         testset = testset[:]
         print(rank, "Local splitting: ", len(trainset), len(valset), len(testset))
-
-        print("Before COMM.Barrier()", flush=True)
         comm.Barrier()
-        print("After COMM.Barrier()", flush=True)
 
         deg = gather_deg(trainset)
         config["pna_deg"] = deg
 
         setnames = ["trainset", "valset", "testset"]
+
+        comm.Barrier()
 
         ## adios
         if args.format == "adios":
@@ -525,6 +285,7 @@ if __name__ == "__main__":
         trainset = AdiosDataset(fname, "trainset", comm, **opt, var_config=var_config)
         valset = AdiosDataset(fname, "valset", comm, **opt, var_config=var_config)
         testset = AdiosDataset(fname, "testset", comm, **opt, var_config=var_config)
+
     elif args.format == "pickle":
         info("Pickle load")
         basedir = os.path.join(
@@ -562,7 +323,11 @@ if __name__ == "__main__":
         os.environ["HYDRAGNN_AGGR_BACKEND"] = "mpi"
         os.environ["HYDRAGNN_USE_ddstore"] = "1"
 
-    (train_loader, val_loader, test_loader,) = hydragnn.preprocess.create_dataloaders(
+    (
+        train_loader,
+        val_loader,
+        test_loader,
+    ) = hydragnn.preprocess.create_dataloaders(
         trainset, valset, testset, config["NeuralNetwork"]["Training"]["batch_size"]
     )
